repository 모델로 KinFragment 적용
