+++ doubleclutch30/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/data/repository/NaverSearchRepository.kt
@@ -1,6 +1,7 @@
 package com.example.architecturestudy.data.repository
 
 import com.example.architecturestudy.data.model.BlogData
+import com.example.architecturestudy.data.model.KinData
 import com.example.architecturestudy.data.model.MovieData
 import com.example.architecturestudy.data.source.remote.NaverSearchRemoteDataSource
 
@@ -9,6 +10,6 @@ interface NaverSearchRepository  {
 
     fun getMovie(keyword : String, success : (MovieData) -> Unit, fail : (Throwable) -> Unit)
     fun getBlog(keyword: String, success: (BlogData) -> Unit, fail: (Throwable) -> Unit)
-
+    fun getKin(keyword: String, success: (KinData) -> Unit, fail: (Throwable) -> Unit)
 
 }
